Locations refactorized.
