+++ app/Http/Controllers/Api/Resource/LocationsController.php
@@ -1,15 +1,13 @@
 <?php
 
-namespace App\Http\Controllers\Api\User;
+namespace App\Http\Controllers\Api\Resource;
 
 use App\Events\Location as LocationEvent;
-use App\Events\UserLocation;
 use App\Http\Controllers\Controller;
-use App\Http\Requests\Api\User\LocationsRequest;
+use App\Http\Requests\Api\Resource\LocationsRequest;
 use App\Http\Resources\Location as LocationResource;
 use App\Repositories\GeofenceRepository;
 use App\Repositories\LocationRepository;
-use App\User;
 use Illuminate\Http\JsonResponse;
 
 class LocationsController extends Controller
@@ -22,27 +20,14 @@ class LocationsController extends Controller
      */
     public function __invoke(LocationsRequest $request)
     {
-        // TODO - Remove unnecessary complexity
-        $user = User::find(auth('api')->user()->id);
-
-        $location = LocationRepository::createLocation(
-            $user,
-            $request->all()
-        );
-
-        $user->update([
+        $model = LocationRepository::resolveLocationModel($request);
+        $location = LocationRepository::createLocation($model, $request->all());
+        $model->update([
             "location" => LocationRepository::parsePoint($request->all())
         ]);
-
-        GeofenceRepository::checkLocationUsingModelGeofences(
-            $user,
-            $location
-        );
-
+        GeofenceRepository::checkLocationUsingModelGeofences($model, $location);
         $payload = new LocationResource($location);
-
-        event(new LocationEvent($payload, $user));
-
+        event(new LocationEvent($payload, $model));
         return response()->json(
             $payload,
             200
